backend/ai-engine/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.3
python-multipart==0.0.6

# Optional: Advanced ML libraries (uncomment if needed)
# scikit-learn==1.4.0
# prophet==1.1.5
# tensorflow==2.15.0
# torch==2.1.2
